refactor(ai-processing-coordinator): move _generate_ai_description implementation into the coordinator (#443)
- Real body of `_generate_ai_description` (OCR + AI call + failure retry path) now lives in `AIProcessingCoordinator._generate_ai_description`.
- Removed the method from EventProcessor (no more forwarding).
- The call inside the coordinator now goes to the local private method.
- `ProcessingContext` no longer carries `generate_ai_description`.

This is the requested next micro-step (option 1) — moving an actual helper implementation into the coordinator while keeping the rest of the helpers on EventProcessor for now.

Part of the AI extraction under Phase B #443.
